+++ cx-portal/src/types/Patterns.ts
@@ -54,6 +54,12 @@ export const Patterns = {
     appLanguage: /^([A-Za-z ])$/,
     pricingInformation: /^([A-Za-z0-9/€ ]){1,15}$/,
   },
+  offerCard: {
+    serviceName: /^([A-Za-z.:_@&0-9' -]){3,20}$/,
+    serviceType: /^([A-Za-z])$/,
+    shortDescriptionEN: /^([a-zA-Z0-9 !?@&#'"()_\-=/*.,;:]){10,120}$/,
+    shortDescriptionDE: /^([a-zA-ZÀ-ÿ0-9 !?@&#'"()_\-=/*.,;:]){10,120}$/,
+  },
   appPage: {
     longDescriptionEN: /^([a-zA-Z0-9 !?@&#'"()[\]_\-+=<>/*.,;:]){10,2000}$/,
     longDescriptionDE: /^([a-zA-ZÀ-ÿ0-9 !?@&#'"()[\]_\-+=<>/*.,;:]){10,2000}$/,