+++ CHANGELOG.md
@@ -7,6 +7,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Navigation base setting
    * Created a new component to support both App Release process & Service Release process intro pages
    * Service Release Steps
+   * Create and Get service apis added
 * Application Request
    * BugFix - Approve/Decline/Retrigger/Cancel Process actions should clear button and update the latest status in both overlay and application list
 * Bugfix - UI fixes in Adminboard